+++ CropsAnytimeAnywhere/Patches/LocationPatcher.cs
@@ -11,7 +11,6 @@ using Pathoschild.Stardew.CropsAnytimeAnywhere.Framework;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Extensions;
-using StardewValley.Locations;
 using xTile.Tiles;
 
 namespace Pathoschild.Stardew.CropsAnytimeAnywhere.Patches
@@ -54,17 +53,8 @@ namespace Pathoschild.Stardew.CropsAnytimeAnywhere.Patches
         public override void Apply(Harmony harmony, IMonitor monitor)
         {
             harmony.Patch(
-                original: this.RequireMethod<GameLocation>(nameof(GameLocation.CanPlantSeedsHere)),
-                postfix: this.GetHarmonyMethod(nameof(LocationPatcher.After_CanPlantSeedsOrTreesHere))
-            );
-
-            harmony.Patch(
-                original: this.RequireMethod<GameLocation>(nameof(GameLocation.CanPlantTreesHere)),
-                postfix: this.GetHarmonyMethod(nameof(LocationPatcher.After_CanPlantSeedsOrTreesHere))
-            );
-            harmony.Patch(
-                original: this.RequireMethod<Town>(nameof(Town.CanPlantTreesHere)), // need to override town separately since it doesn't check base
-                postfix: this.GetHarmonyMethod(nameof(LocationPatcher.After_CanPlantSeedsOrTreesHere))
+                original: typeof(GameLocation).GetMethod(nameof(GameLocation.CheckItemPlantRules)) ?? throw new InvalidOperationException($"Can't find method {nameof(GameLocation.CheckItemPlantRules)}"),
+                prefix: this.GetHarmonyMethod(nameof(LocationPatcher.Before_CheckItemPlantRules))
             );
 
             harmony.Patch(
@@ -88,26 +78,27 @@ namespace Pathoschild.Stardew.CropsAnytimeAnywhere.Patches
         /****
         ** Patches
         ****/
-        /// <summary>A method called via Harmony after <see cref="GameLocation.CanPlantSeedsHere"/> or <see cref="GameLocation.CanPlantTreesHere"/>.</summary>
-        /// <param name="__instance">The farm instance.</param>
-        /// <param name="__result">The return value to use for the method.</param>
-        private static void After_CanPlantSeedsOrTreesHere(GameLocation __instance, ref bool __result)
+        /// <summary>A method called via Harmony before <see cref="GameLocation.CheckItemPlantRules"/>.</summary>
+        /// <param name="__instance">The location instance.</param>
+        /// <param name="defaultAllowed">The result to return when no rules apply, or the selected rule uses <see cref="F:StardewValley.GameData.PlantableResult.Default" />.</param>
+        [SuppressMessage("ReSharper", "RedundantAssignment", Justification = "Matches original code code")]
+        private static void Before_CheckItemPlantRules(GameLocation __instance, ref bool defaultAllowed)
         {
-            if (!__result && LocationPatcher.Config.TryGetForLocation(__instance, out PerLocationConfig? config) && config.GrowCrops)
-                __result = true;
+            if (!defaultAllowed && LocationPatcher.Config.TryGetForLocation(__instance, out PerLocationConfig? config) && config.GrowCrops)
+                defaultAllowed = true;
         }
 
         /// <summary>A method called via Harmony after <see cref="GameLocation.SeedsIgnoreSeasonsHere"/>.</summary>
-        /// <param name="__instance">The farm instance.</param>
+        /// <param name="__instance">The location instance.</param>
         /// <param name="__result">The return value to use for the method.</param>
         private static void After_SeedsIgnoreSeasonsHere(GameLocation __instance, ref bool __result)
         {
-            if (!__result && LocationPatcher.Config.TryGetForLocation(__instance, out PerLocationConfig? config) && config.GrowCrops && config.GrowCropsOutOfSeason && !LocationPatcher.IsGameClearingTilledDirt())
+            if (!__result && LocationPatcher.Config.TryGetForLocation(__instance, out PerLocationConfig? config) && config is { GrowCrops: true, GrowCropsOutOfSeason: true } && !LocationPatcher.IsGameClearingTilledDirt())
                 __result = true;
         }
 
         /// <summary>A method called via Harmony after <see cref="GameLocation.doesTileHaveProperty"/>.</summary>
-        /// <param name="__instance">The farm instance.</param>
+        /// <param name="__instance">The location instance.</param>
         /// <param name="xTile">The x-coordinate of the map tile.</param>
         /// <param name="yTile">The y-coordinate of the map tile.</param>
         /// <param name="propertyName">The property name to match.</param>
@@ -172,7 +163,7 @@ namespace Pathoschild.Stardew.CropsAnytimeAnywhere.Patches
         /// <summary>Get the value of a tile or tile index property.</summary>
         /// <param name="tile">The tile to check.</param>
         /// <param name="name">The property name.</param>
-        /// <remarks>Derived from <see cref="GameLocation.doesTileHaveProperty(int, int, string, string)"/> with optimizations.</remarks>
+        /// <remarks>Derived from <see cref="GameLocation.doesTileHaveProperty"/> with optimizations.</remarks>
         private static string? GetProperty(Tile tile, string name)
         {
             if (tile.TileIndexProperties?.TryGetValue(name, out string? value) is true)